Cli Github OAuth Setup
This uses githubs device flow to get gh creds in the cli and provision a repository automagically.  The advantage here is we can also auto-setup deploy keys ensuring proper git configuration for console installations at least for github users.

An extensible scm provider interface is defined as well for when we want to add gitlab+bitbucket support.